Enable feed HEAD requests
