+++ PodcastRewind/Pages/Details.cshtml.cs
@@ -45,7 +45,7 @@ public class DetailsModel : PageModel
 
         ScheduledFeed = FeedRewindData.GetScheduledFeed();
         RewindFeedId = id.Value;
-        RewindFeedUrl = Url.ActionLink("GetFeed", "Feed", new { id })!;
+        RewindFeedUrl = Url.ActionLink("Get", "Feed", new { id })!;
         ApplePodcastSubscribeUrl = Url.ActionLink("GetFeed", "Feed", new { id }, "podcast")!;
         PodcastImageUrl = RewoundFeed.ImageUrl?.ToString() ?? "";
         OriginalPodcastLink = FeedRewindData.OriginalFeedLink?.AbsoluteUri;
